# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: "Описание спецификации всех API методов, используемых в online diffusion\
    \ сервисе"
  title: Online diffusion API
  version: 0.0.1
  x-logo:
    url: https://i.imgur.com/1kUh580.jpg
...
servers:
- description: "Локальный сервер, актуально для запуска на своем компьютере"
  url: http://localhost:8080
  x-last-modified: 1698183338768
security:
- JWTAuth: []
...
paths:
...
  /api/v1/generator:
    get:
      description: ""
      operationId: get-generator
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models'
          description: все хорошо присылаются модели
      security:
      - JWTAuth: []
      summary: Получаем список моделей для генератора
      tags:
      - Generator
      x-accepts: application/json
      x-tags:
      - tag: Generator
    post:
      description: ""
      operationId: post-generator
      parameters:
      - description: prompt для создания фото
        explode: true
        in: query
        name: prompt
        required: true
        schema:
          example: cool looking panda
          type: string
        style: form
      - description: anti-prompt для создания фото
        explode: true
        in: query
        name: anti-prompt
        required: true
        schema:
          example: ugly face
          type: string
        style: form
      - description: Seed для генерации нейросети. Оставить пустым для случайного
        explode: true
        in: query
        name: seed
        required: false
        schema:
          example: 123456
          format: int64
          type: integer
        style: form
      - description: "Название модели, которая будет использоваться для генерации"
        explode: true
        in: query
        name: modelName
        required: true
        schema:
          example: anime-diffusion
          type: string
        style: form
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generationToken'
          description: отправлено на генерацию
        "400":
          description: "если не заполнены prompt, anti-prompt или model_name"
      security:
      - JWTAuth: []
      summary: Начинаем генерацию изображения
      tags:
      - Generator
      x-accepts: application/json
      x-tags:
      - tag: Generator
    trace:
      description: ""
      operationId: trace-generator
      parameters:
      - description: "Токен, по которому можно узнать об статусе генерируемого изо\
          бражения"
        explode: true
        in: query
        name: generationToken
        required: true
        schema:
          example: "123"
          type: string
        style: form
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
          description: нейронная сеть успешно сгенерировала изображение и в присланном
            Image будет ссылка на него
        "202":
          description: "генерация изображения продолжается и еще не была завершена\
            , но и не сломалась"
        "404":
          description: такого generation_token нет на сервере в принципе
        "500":
          description: во время генерации изображения случилась ошибка и оно было
            потеряно безвозвратно :(
      security:
      - JWTAuth: []
      summary: Получаем сделанное изображение в цикле
      tags:
      - Generator
      x-accepts: application/json
      x-tags:
      - tag: Generator
...
components:
...
  schemas:
...
    Image:
      description: Объект с информацией об фотографии
      example:
        seed: seed
        authorName: authorName
        authorAvatarUrl: authorAvatarUrl
        photoId: 0.8008281904610115
        anti-prompt: anti-prompt
        model: model
        prompt: prompt
      properties:
        photoId:
          description: Id фотографии
          type: number
        authorName:
          description: Имя автора
          type: string
        authorAvatarUrl:
          description: URL аватара автора
          type: string
        prompt:
          description: prompt использовшийся для создания изображения
          type: string
        anti-prompt:
          description: anti-prompt использовшийся для создания изображения
          type: string
        seed:
          description: "seed, использовшийся для создания изображения"
          type: string
        model:
          description: "Модель, использувшееся для создания изображения"
          type: string
      required:
      - anti-prompt
      - authorAvatarUrl
      - authorName
      - model
      - photoId
      - prompt
      - seed
      type: object
    generationToken:
      description: "Токен, назначаемый сессии генерации изображения"
      type: string
    models:
      description: "Все модели, которые можно использовать для генерации"
      items:
        description: Модель для генерации
        type: string
      type: array
...
  securitySchemes:
    JWTAuth:
      bearerFormat: JWT
      description: "Example: \n> Authorization: Bearer <token>"
      scheme: bearer
      type: http
      x-last-modified: 1698183246392